feat(dashboard): refactor dashboard (1)
move utility functions into utils directory
